requirements.txt
astroid==3.2.4
colorama==0.4.6
comtypes==1.4.6
dill==0.3.8
isort==5.13.2
jsonlint==0.1
mccabe==0.7.0
MouseInfo==0.1.3
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
platformdirs==4.2.2
psutil==6.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pylint==3.2.6
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pystray==0.19.5
python-dotenv==1.0.1
pytweening==1.2.0
pywin32==306
pywinauto==0.6.8
six==1.16.0
tomli==2.0.1
tomlkit==0.13.0
typing_extensions==4.12.2
